introduce datamodel and API doc
