+++ docs/api-datamodel.md
@@ -1,0 +1,2 @@
+# Data Model
+::: sv_simpleparser.datamodel
