+++ docs/api-parser.md
@@ -1,0 +1,2 @@
+# Parser
+::: sv_simpleparser.parser
