+++ docs/api.md
@@ -1,1 +1,5 @@
+# Application Programming Interface (API)
+
 ::: sv_simpleparser
+    options:
+        members: []
